+++ src/cli/index.ts
@@ -3,7 +3,7 @@
 import { parseArgs } from 'node:util';
 import { repl } from './repl.js';
 import { run } from './run.js';
-import { test } from './test.js';
+import { test } from './test/test.js';
 
 const args = parseArgs({
     allowPositionals: true,
